Правки в функции модального окна
