+++ index.html
@@ -227,6 +227,5 @@
     </section>
   </template>
   <script src="js/main.js" type="module"></script>
-  <script src="js/functions.js"></script>
 </body>
 </html>
