+++ js/display-big-picture.js
@@ -1,5 +1,3 @@
-import { showModal } from './modal-show.js';
-
 const bigPicture = document.querySelector('.big-picture');
 const bigPictureImage = bigPicture.querySelector('.big-picture__img').querySelector('img');
 const likesCount = bigPicture.querySelector('.likes-count');
@@ -7,19 +5,47 @@ const commentsCount = bigPicture.querySelector('.comments-count');
 const commentsList = bigPicture.querySelector('.social__comments');
 const socialComment = bigPicture.querySelector('.social__comment');
 const socialCaption = bigPicture.querySelector('.social__caption');
+const closeButton = document.querySelector('.big-picture__cancel');
+
+const hideElements = () => {
+  document.querySelector('.social__comment-count').classList.add('hidden');
+  document.querySelector('.comments-loader').classList.add('hidden');
+};
+
+const onDocumentKeydown = (evt) => {
+  if (evt.key === 'Escape' && !evt.target.closest('.social__footer-text')) {
+    evt.preventDefault();
+    closeModal();
+  }
+};
+
+const onButtonCloseClick = () => closeModal ();
+
+// Использую объявление функции через function для всплытия - для функции onModalEscapeKeydown
+const showModal = () => {
+  bigPicture.classList.remove('hidden');
+  document.body.classList.add('modal-open');
+  document.addEventListener('keydown', onDocumentKeydown);
+  closeButton.addEventListener('click', onButtonCloseClick);
+};
+
+function closeModal () {
+  bigPicture.classList.add('hidden');
+  document.body.classList.remove('modal-open');
+  document.removeEventListener('keydown', onDocumentKeydown);
+  closeButton.removeEventListener('click', onButtonCloseClick);
+}
 
 const createComment = (item) => {
   const comment = socialComment.cloneNode(true);
   const avatar = comment.querySelector('.social__picture');
-  const text = comment.querySelector('.social__text');
   avatar.src = item.avatar;
   avatar.alt = item.name;
-  text.textContent = item.message;
+  comment.querySelector('.social__text').textContent = item.message;
   return comment;
 };
 
 const createCommentsList = ({ comments }) => {
-  commentsList.innerHTML = '';
   comments.forEach((comment) => commentsList.append(createComment(comment)));
 };
 
@@ -33,8 +59,12 @@ const fillBigPicture = (data) => {
 };
 
 const displayBigPicture = (data) => {
+  commentsList.innerHTML = '';
+  hideElements();
   fillBigPicture(data);
   showModal();
 };
 
 export { displayBigPicture };
+
+// const displayBigPicture = () => {};
